DLCs_Main: Moved the DLC_*.h data files to a subdirectory.
